// frontend-libreria-don-hector/src/features/ConfigAlerts/types/config.types.ts
export interface Product {
    id_producto: number;
    nombre: string;
    descripcion: string;
    codigo_producto: string;
    categoria: string;
    precio_compra: number;
    precio_venta: number;
    stock: number;
    imagen: string;
    es_libro: boolean;
    autor: string | null;
    fecha_lanzamiento: string | null;
    stock_minimo: number;
  }
  
  export interface ProductUpdateData {
    id_producto: number;
    nombre: string;
    descripcion: string;
    codigo_producto: string;
    categoria: string;
    precio_compra: number;
    precio_venta: number;
    stock: number;
    imagen: string;
    es_libro: boolean;
    autor: string | null;
    fecha_lanzamiento: string | null;
    stock_minimo: number;
  }
  
  export interface BulkUpdateRequest {
    products: ProductUpdateData[];
  }
  
  export interface ProductResponse {
    status: string;
    message: string;
    data: {
      products: Product[];
    };
  }